requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# Multi-Agent System (use latest compatible versions)
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# API Clients
tavily-python>=0.3.3
openai>=1.50.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
